Modified(Battery implementation): Modified the Battery class for the drone and Robot such that the entities are now under the control of their respective battery capacity to be able to update their movement and also deliver the Package
- If drone or Robot has zero battery, they move to Idle state and do not
drain an further battery
- The delivery simulation reschedules the packages once the Robot's or Drone's
Battery is completely drained
- Package is reassigned but not dropped after Drone or Robot dies of low
battery_level

+++ project/include/Battery.h
@@ -58,7 +58,7 @@ public:
 	 * @brief this function controls the battery level, changing it according to the value passed by the user
 	 * @param delta is the amount by how much the battery level should be changed
 	 */
-	void change_level(int delta);
+	void change_level(float delta);
 
 	/**
 	 * @brief returns the model of the Battery in string format
